requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.59.0
imageio==2.37.0
importlib_resources==6.5.2
# Editable install with no version control (IXE-package==0.1.0)
./
kiwisolver==1.4.7
lazy_loader==0.4
matplotlib==3.9.4
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scikit-image==0.24.0
scipy==1.13.1
six==1.17.0
tifffile==2024.8.30
zipp==3.23.0
